Fix done to the collision, not working 100% but enough to push

+++ Breakout/Brick.cpp
@@ -26,7 +26,7 @@ namespace Logic
 
 	}
 
-	void Brick::update(double _dt)
+	void Brick::update(float _dt)
 	{
 
 	}
