+++ Breakout/Object3D.h
@@ -12,7 +12,7 @@ namespace Logic
 		Object3D()
 		{
 		}
-		virtual void update(double _dt) = 0;
+		virtual void update(float _dt) = 0;
 
 		int getModelID()		{ return modelID;	};
 		int getTextureID()		{ return textureID;	};
